ADM-688:[frontend] refactor: Move button group to component.
