+++ frontend/src/components/Metrics/ReportStep/index.tsx
@@ -1,40 +1,28 @@
 import React, { useEffect, useLayoutEffect, useState } from 'react'
 import { useGenerateReportEffect } from '@src/hooks/useGenerateReportEffect'
 import { useAppSelector } from '@src/hooks'
-import { selectConfig, selectJiraColumns, selectMetrics } from '@src/context/config/configSlice'
-import { BOARD_METRICS, CALENDAR, DORA_METRICS, MESSAGE, REQUIRED_DATA, TIPS } from '@src/constants/resources'
-import { COMMON_BUTTONS, DOWNLOAD_TYPES } from '@src/constants/commons'
-import { BoardReportRequestDTO, CSVReportRequestDTO, ReportRequestDTO } from '@src/clients/report/dto/request'
+import { selectConfig, selectJiraColumns } from '@src/context/config/configSlice'
+import { BOARD_METRICS, CALENDAR, DORA_METRICS, MESSAGE, REQUIRED_DATA } from '@src/constants/resources'
+import { BoardReportRequestDTO, ReportRequestDTO } from '@src/clients/report/dto/request'
 import { IPipelineConfig, selectMetricsContent } from '@src/context/Metrics/metricsSlice'
 import dayjs from 'dayjs'
-import { BackButton, SaveButton } from '@src/components/Metrics/MetricsStepper/style'
-import { useExportCsvEffect } from '@src/hooks/useExportCsvEffect'
-import { backStep, selectTimeStamp } from '@src/context/stepper/StepperSlice'
-import { useAppDispatch } from '@src/hooks/useAppDispatch'
-import {
-  StyledButtonGroup,
-  StyledErrorNotification,
-  StyledExportButton,
-  StyledMetricsSection,
-  StyledSpacing,
-} from '@src/components/Metrics/ReportStep/style'
+import { selectTimeStamp } from '@src/context/stepper/StepperSlice'
+import { StyledErrorNotification, StyledMetricsSection, StyledSpacing } from '@src/components/Metrics/ReportStep/style'
 import { ErrorNotification } from '@src/components/ErrorNotification'
 import { useNavigate } from 'react-router-dom'
-import { ExpiredDialog } from '@src/components/Metrics/ReportStep/ExpiredDialog'
 import { filterAndMapCycleTimeSettings, getJiraBoardToken } from '@src/utils/util'
 import { useNotificationLayoutEffectInterface } from '@src/hooks/useNotificationLayoutEffect'
 import { ROUTE } from '@src/constants/router'
-import { Tooltip } from '@mui/material'
-import SaveAltIcon from '@mui/icons-material/SaveAlt'
 import { ReportTitle } from '@src/components/Common/ReportGrid/ReportTitle/ReportTitle'
 import { ReportGrid } from '@src/components/Common/ReportGrid'
+import { ReportButtonGroup } from '@src/components/Metrics/ReportButtonGroup'
 
 export interface ReportStepProps {
   notification: useNotificationLayoutEffectInterface
   handleSave: () => void
 }
+
 const ReportStep = ({ notification, handleSave }: ReportStepProps) => {
-  const dispatch = useAppDispatch()
   const navigate = useNavigate()
   const {
     isServerError,
@@ -90,7 +78,7 @@ const ReportStep = ({ notification, handleSave }: ReportStepProps) => {
       },
     },
   } = useGenerateReportEffect()
-  const { fetchExportData, errorMessage: csvErrorMsg, isExpired } = useExportCsvEffect()
+
   const [exportValidityTimeMin] = useState<number | undefined>(undefined)
   const csvTimeStamp = useAppSelector(selectTimeStamp)
   const configData = useAppSelector(selectConfig)
@@ -109,38 +97,13 @@ const ReportStep = ({ notification, handleSave }: ReportStepProps) => {
   const { board, pipelineTool, sourceControl } = configData
   const { token, type, site, projectKey, boardId, email } = board.config
   const { startDate, endDate } = dateRange
-  const requiredData = useAppSelector(selectMetrics)
-  const isShowExportBoardButton =
-    requiredData.includes(REQUIRED_DATA.VELOCITY) ||
-    requiredData.includes(REQUIRED_DATA.CYCLE_TIME) ||
-    requiredData.includes(REQUIRED_DATA.CLASSIFICATION)
-  const isShowExportPipelineButton =
-    requiredData.includes(REQUIRED_DATA.DEPLOYMENT_FREQUENCY) ||
-    requiredData.includes(REQUIRED_DATA.CHANGE_FAILURE_RATE) ||
-    requiredData.includes(REQUIRED_DATA.LEAD_TIME_FOR_CHANGES) ||
-    requiredData.includes(REQUIRED_DATA.MEAN_TIME_TO_RECOVERY)
   const { updateProps } = notification
-  const [errorMessage, setErrorMessage] = useState([reportErrorMsg, csvErrorMsg])
+  const [errorMessage, setErrorMessage] = useState<string>()
   const jiraToken = getJiraBoardToken(token, email)
   const jiraColumns = useAppSelector(selectJiraColumns)
   const jiraColumnsWithValue = jiraColumns?.map(
     (obj: { key: string; value: { name: string; statuses: string[] } }) => obj.value
   )
-
-  const handleDownload = (dataType: DOWNLOAD_TYPES, startDate: string | null, endDate: string | null) => {
-    fetchExportData(getExportCSV(dataType, startDate, endDate))
-  }
-
-  const getExportCSV = (
-    dataType: DOWNLOAD_TYPES,
-    startDate: string | null,
-    endDate: string | null
-  ): CSVReportRequestDTO => ({
-    dataType: dataType,
-    csvTimeStamp: csvTimeStamp,
-    startDate: startDate ?? '',
-    endDate: endDate ?? '',
-  })
 
   const getPipelineConfig = (pipelineConfigs: IPipelineConfig[]) => {
     if (!pipelineConfigs[0].organization && pipelineConfigs.length === 1) {
@@ -218,23 +181,6 @@ const ReportStep = ({ notification, handleSave }: ReportStepProps) => {
     }
   }
 
-  const handleBack = () => {
-    dispatch(backStep())
-  }
-
-  const handleErrorNotification = () => {
-    {
-      return errorMessage.map((message: string) => {
-        if (message === '') return
-        return (
-          <StyledErrorNotification key={message}>
-            <ErrorNotification message={message} />
-          </StyledErrorNotification>
-        )
-      })
-    }
-  }
-
   useLayoutEffect(() => {
     exportValidityTimeMin &&
       updateProps?.({
@@ -274,8 +220,8 @@ const ReportStep = ({ notification, handleSave }: ReportStepProps) => {
   }, [])
 
   useEffect(() => {
-    setErrorMessage([reportErrorMsg, csvErrorMsg])
-  }, [reportErrorMsg, csvErrorMsg])
+    setErrorMessage(reportErrorMsg)
+  }, [reportErrorMsg])
 
   useEffect(() => {
     return () => {
@@ -401,7 +347,11 @@ const ReportStep = ({ notification, handleSave }: ReportStepProps) => {
       ) : (
         <>
           {/*{startDate && endDate && <CollectionDuration startDate={startDate} endDate={endDate} />}*/}
-          {handleErrorNotification()}
+          {errorMessage && (
+            <StyledErrorNotification>
+              <ErrorNotification message={errorMessage} />
+            </StyledErrorNotification>
+          )}
           <>
             {reportData && (
               <StyledMetricsSection>
@@ -420,32 +370,15 @@ const ReportStep = ({ notification, handleSave }: ReportStepProps) => {
             )}
           </>
 
-          <StyledButtonGroup>
-            <Tooltip title={TIPS.SAVE_CONFIG} placement={'right'}>
-              <SaveButton variant='text' onClick={handleSave} startIcon={<SaveAltIcon />}>
-                {COMMON_BUTTONS.SAVE}
-              </SaveButton>
-            </Tooltip>
-            <div>
-              <BackButton onClick={handleBack} variant='outlined'>
-                {COMMON_BUTTONS.BACK}
-              </BackButton>
-              <StyledExportButton onClick={() => handleDownload(DOWNLOAD_TYPES.METRICS, startDate, endDate)}>
-                {COMMON_BUTTONS.EXPORT_METRIC_DATA}
-              </StyledExportButton>
-              {isShowExportBoardButton && (
-                <StyledExportButton onClick={() => handleDownload(DOWNLOAD_TYPES.BOARD, startDate, endDate)}>
-                  {COMMON_BUTTONS.EXPORT_BOARD_DATA}
-                </StyledExportButton>
-              )}
-              {isShowExportPipelineButton && (
-                <StyledExportButton onClick={() => handleDownload(DOWNLOAD_TYPES.PIPELINE, startDate, endDate)}>
-                  {COMMON_BUTTONS.EXPORT_PIPELINE_DATA}
-                </StyledExportButton>
-              )}
-            </div>
-          </StyledButtonGroup>
-          {<ExpiredDialog isExpired={isExpired} handleOk={handleBack} />}
+          <ReportButtonGroup
+            handleSave={() => handleSave()}
+            csvTimeStamp={csvTimeStamp}
+            startDate={startDate}
+            endDate={endDate}
+            setErrorMessage={(message) => {
+              setErrorMessage(message)
+            }}
+          />
         </>
       )}
     </>
