+++ frontend/src/components/Metrics/ReportStep/style.tsx
@@ -15,25 +15,3 @@ export const StyledMetricsSection = styled('div')({
 export const StyledSpacing = styled('div')({
   height: '1.5rem',
 })
-
-export const StyledButtonGroup = styled('div')({
-  boxSizing: 'border-box',
-  display: 'flex',
-  textAlign: 'center',
-  margin: '0 auto',
-  justifyContent: 'space-between',
-  width: '100%',
-  paddingTop: '2rem',
-})
-
-export const StyledExportButton = styled(Button)({
-  ...basicButtonStyle,
-  width: '12rem',
-  backgroundColor: theme.main.backgroundColor,
-  color: theme.main.color,
-  '&:hover': {
-    ...basicButtonStyle,
-    backgroundColor: theme.main.backgroundColor,
-    color: theme.main.color,
-  },
-})
